UI Cleanup
Cleaned up most of the controllers and the application delegate, and
reconfigured the way startup is handled. Also, removed sublcasses of
UINavigationView. Subclassing UINavigationView is not permitted and
will get your app rejected from the app store. Removed some dead code
an commented-out code. As a fortuitous consequence, the app now works
in iOS 5.1.